ext-source/docs-notebooks/voyageai/apps/rag-with-mongodb/requirements.txt
voyageai
numpy
pymongo
langchain-community
langchain-text-splitters
pypdf
python-dotenv
anthropic
openai